+++ NEWS.md
@@ -1,3 +1,7 @@
 # chkMocks 0.1.0
 
+* Added a `plotZymoDefault` function to the package.
+* Added a `assignTaxonomyCustomMock` function to the package.
+* Added a `compare2theorectical` function to the package.
+* Added a `checkZymoBiomics` function to the package.
 * Added a `NEWS.md` file to track changes to the package.